refactor(platform): standardize on direct @platform/platform import path
The @platform barrel (src/platform/index.ts) was bypassed by ~95% of
importers: only 4 files used it while 90 imported @platform/platform
directly, including central call sites like initPlatform.ts and the
node hosts. The documented convention (CLAUDE.md/AGENTS.md pointed at
the barrel) no longer matched actual usage, and the unused barrel was
dead weight with no logic beyond re-exporting platform.ts.

Standardize on the already-dominant direct path: convert the 4 barrel
importers to @platform/platform, delete the now-fully-unused barrel
file, drop the bare "@platform" tsconfig path entry (all three
tsconfig files share a single derivation for esbuild/Vite/Vitest
aliases via scripts/aliases.mjs), and update the CLAUDE.md/AGENTS.md
guidance that recommended the barrel.

Fixes #7423.
